+++ pyam/utils.py
@@ -1,11 +1,14 @@
+import importlib.metadata
 import itertools
 import logging
 import re
 import string
+import warnings
 from pathlib import Path
 
 import dateutil
 import numpy as np
+import packaging.version
 import pandas as pd
 from pandas.api.types import is_list_like
 
@@ -94,13 +97,38 @@ def write_sheet(writer, name, df, index=False):
         writer.sheets[name].set_column(i, i, width)  # assumes xlsxwriter as engine
 
 
+def get_excel_file_with_kwargs(path, **kwargs):
+    """Return a `pandas.ExcelFile` and a dict of unused kwargs.
+
+    When reading an Excel file, this function finds keyword arguments that
+    should be passed to `pandas.ExcelFile`, and returns a `pandas.ExcelFile`
+    instance along with the remaining keyword arguments (which presumably
+    will be used for other purposes by the calling function).
+    """
+    EXCEL_FILE_KWS = ('engine', 'storage_options', 'engine_kwargs')
+    kwargs = kwargs.copy()
+    excel_file_kwargs = {
+        k: kwargs.pop(k) for k in EXCEL_FILE_KWS if k in kwargs
+    }
+    # TODO remove when bumping minimum pandas dependency to >= 2.2
+    if "engine_kwargs" in excel_file_kwargs and packaging.version.parse(
+        importlib.metadata.version("pandas")
+    ) < packaging.version.parse("2.2.0"):
+        warnings.warn(
+            "pandas < 2.2.0 has inconsistent support for `engine_kwargs`. "
+            "Using it is likely to result in an exception."
+        )
+    return pd.ExcelFile(path, **excel_file_kwargs), kwargs
+
+
 def read_pandas(path, sheet_name=["data*", "Data*"], *args, **kwargs):
     """Read a file and return a pandas.DataFrame"""
 
     if isinstance(path, Path) and path.suffix == ".csv":
         return pd.read_csv(path, *args, **kwargs)
 
-    with pd.ExcelFile(path) as xl:
+    xlfile, kwargs = get_excel_file_with_kwargs(path, **kwargs)
+    with xlfile as xl:
         # reading multiple sheets
         sheet_names = pd.Series(xl.sheet_names)
         if len(sheet_names) > 1: